requirements.txt
fastapi==0.79.0
uvicorn[standard]==0.18.2
celery==5.2.7
redis==4.3.4
python-dotenv==1.0.0
